+++ src/intelli/core.h
@@ -8,6 +8,10 @@ namespace intelli
 
 GT_INTELLI_EXPORT void initModule();
 
+GT_INTELLI_EXPORT void registerDefaultDataTypes();
+
+GT_INTELLI_EXPORT void registerDefaultNodes();
+
 } // namespace intelli
 
 #endif // INTELLI_CORE_H